requirements.txt
Django==5.2.9
psycopg2-binary==2.9.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1